+++ src/checks/indentPref.js
@@ -6,8 +6,8 @@
  * @returns {boolean} true if # of spaces correct, false if not
  */
 var indentPref = function() {
-  var spaces = this.state.conf
-  var context = this.state.context
+	var spaces = this.state.conf
+	var context = this.state.context
 
 	if ( typeof spaces !== 'number' ) { return }
 
@@ -17,9 +17,9 @@ var indentPref = function() {
 		indentCorrect = false
 	}
 
-  // if spaces === 2 and context === 1.5 (meaning 1.5 levels... or 3 spaces)
-  // then the index for the warning msg is 2 * 1.5, or 3 spaces
-  var index = spaces * context
+	// if spaces === 2 and context === 1.5 (meaning 1.5 levels... or 3 spaces)
+	// then the index for the warning msg is 2 * 1.5, or 3 spaces
+	var index = spaces * context
 
 	if ( indentCorrect === false ) {
 		this.msg( 'incorrect # of spaces for indent, use ' + spaces, index )