+++ src/checks/noImportant.js
@@ -8,7 +8,7 @@
  */
 var noImportant = function( line ) {
 	var important = false
-  var index = line.indexOf( '!important' )
+	var index = line.indexOf( '!important' )
 
 	// return true if border|outline is followed by a 0
 	if ( index !== -1 ) {