+++ src/utils/msg.js
@@ -1,7 +1,5 @@
 'use strict'
 
-var chalk = require( 'chalk' )
-
 
 /**
  * @description basically just sets the severity and routes output to the reporter
@@ -12,7 +10,7 @@ var chalk = require( 'chalk' )
 var msg = function( str, column ) {
 	// determine which group the msg belongs to
 	var arr = this.state.severity === 'Warning' ? this.cache.warnings : this.cache.errs
-  this.cache.col = column
+	this.cache.col = column
 
 	// push the final output
 	return arr.push( this.reporter( str ) )