+++ changelog.md
@@ -4,7 +4,7 @@ adds column number for errors to output
 
 improves default reporter output
 
-string and whitespacing is now much improved: https://github.com/rossPatton/stylint/issues/302
+string and whitespace parsing is now much improved: https://github.com/rossPatton/stylint/issues/302
 
 set preferGlobal to false: https://github.com/rossPatton/stylint/issues/278